updating /posts/ routes to be /blog/ routes
