+++ src/app/(frontend)/[center]/blog/page.tsx
@@ -5,38 +5,29 @@ import { Pagination } from '@/components/Pagination'
 import { PostCollection } from '@/components/PostCollection'
 import { POSTS_LIMIT } from '@/utilities/constants'
 import configPromise from '@payload-config'
-import { notFound } from 'next/navigation'
 import { getPayload } from 'payload'
-import { PostsSort } from '../../posts-sort'
-import { PostsTags } from '../../posts-tags'
-
-export const revalidate = 600
+import { PostsSort } from './posts-sort'
+import { PostsTags } from './posts-tags'
 
 type Args = {
   params: Promise<{
-    center: string
-    pageNumber: string
+    center: string[]
   }>
   searchParams: Promise<{ [key: string]: string }>
 }
 
-export default async function Page({ params: paramsPromise, searchParams }: Args) {
+export default async function Page({ params, searchParams }: Args) {
+  const { center } = await params
   const resolvedSearchParams = await searchParams
   const sort = resolvedSearchParams?.sort || '-publishedAt'
   const selectedTag = resolvedSearchParams?.tags
 
-  const { center, pageNumber } = await paramsPromise
   const payload = await getPayload({ config: configPromise })
-
-  const sanitizedPageNumber = Number(pageNumber)
-
-  if (!Number.isInteger(sanitizedPageNumber)) notFound()
 
   const posts = await payload.find({
     collection: 'posts',
     depth: 2,
     limit: POSTS_LIMIT,
-    page: sanitizedPageNumber,
     where: {
       'tenant.slug': {
         equals: center,
@@ -66,13 +57,13 @@ export default async function Page({ params: paramsPromise, searchParams }: Args
 
   return (
     <div className="py-6">
-      <div className="container mb-16 flex flex-col-reverse md:flex-row flex-1 gap-6">
+      <div className="container md:max-lg:max-w-5xl mb-16 flex flex-col-reverse md:flex-row flex-1 gap-6">
         <div className="grow">
           <PostCollection posts={posts.docs} />
         </div>
 
         {/* Sorting and filters */}
-        <div className="flex flex-col gap-4 shrink-0 justify-between md:justify-start md:w-[300px]">
+        <div className="flex flex-col gap-4 shrink-0 justify-between md:justify-start md:w-[240px] lg:w-[300px]">
           <PostsSort initialSort={sort} />
           {tags.docs.length > 1 && <PostsTags tags={tags.docs} />}
         </div>
@@ -99,7 +90,7 @@ export default async function Page({ params: paramsPromise, searchParams }: Args
 
 export async function generateMetadata({ params }: Args): Promise<Metadata> {
   const payload = await getPayload({ config: configPromise })
-  const { center, pageNumber } = await params
+  const { center } = await params
   const tenant = await payload.find({
     collection: 'tenants',
     select: {
@@ -113,27 +104,10 @@ export async function generateMetadata({ params }: Args): Promise<Metadata> {
   })
   if (tenant.docs.length < 1) {
     return {
-      title: `Avalanche Center Posts Page ${pageNumber || ''}`,
+      title: `Avalanche Center Blog`,
     }
   }
   return {
-    title: `${tenant.docs[0].name} - Posts Page ${pageNumber || ''}`,
+    title: `${tenant.docs[0].name} - Blog`,
   }
-}
-
-export async function generateStaticParams() {
-  const payload = await getPayload({ config: configPromise })
-  const { totalDocs } = await payload.count({
-    collection: 'posts',
-  })
-
-  const totalPages = Math.ceil(totalDocs / 10)
-
-  const pages: { pageNumber: string }[] = []
-
-  for (let i = 1; i <= totalPages; i++) {
-    pages.push({ pageNumber: String(i) })
-  }
-
-  return pages
 }
