+++ src/collections/Posts/hooks/revalidatePost.ts
@@ -12,7 +12,7 @@ export const revalidatePost: CollectionAfterChangeHook<Post> = ({
   if (!context.disableRevalidate) {
     if (doc._status === 'published') {
       // TODO: revalidate path based on tenant
-      const path = `/posts/${doc.slug}`
+      const path = `/blog/${doc.slug}`
 
       payload.logger.info(`Revalidating post at path: ${path}`)
 
@@ -22,7 +22,7 @@ export const revalidatePost: CollectionAfterChangeHook<Post> = ({
 
     // If the post was previously published, we need to revalidate the old path
     if (previousDoc._status === 'published' && doc._status !== 'published') {
-      const oldPath = `/posts/${previousDoc.slug}`
+      const oldPath = `/blog/${previousDoc.slug}`
 
       payload.logger.info(`Revalidating old post at path: ${oldPath}`)
 
@@ -35,7 +35,7 @@ export const revalidatePost: CollectionAfterChangeHook<Post> = ({
 
 export const revalidateDelete: CollectionAfterDeleteHook<Post> = ({ doc, req: { context } }) => {
   if (!context.disableRevalidate) {
-    const path = `/posts/${doc?.slug}`
+    const path = `/blog/${doc?.slug}`
 
     revalidatePath(path)
     revalidateTag('posts-sitemap')
