+++ src/components/Card/index.tsx
@@ -23,7 +23,7 @@ export const Card = (props: {
 
   const titleToUse = titleFromProps || title
   const sanitizedDescription = description?.replace(/\s/g, ' ') // replace non-breaking space with white space
-  const href = `/${relationTo}/${slug}`
+  const href = `/${relationTo === 'posts' ? 'blog' : relationTo}/${slug}`
 
   return (
     <article
